MI-326: Implement conditional shim and fix incorrect main path
