+++ packages/vite-plugin-handler/src/index.ts
@@ -1,3 +1,3 @@
 export { handlerBundle } from './lib/handler-bundle.js';
 export type { HandlerBundleOptions } from './lib/handler-bundle.js';
-export type { ConditionalShim } from './lib/plugins.js';
+export type { ConditionalShim } from './lib/shim.js';
